requirements.txt
numpy
pyqubo
dwave-ocean-sdk
matplotlib